+++ crates/codegen/src/backend/evm/stack_schedule.rs
